AWS SDK for iOS 1.6.1

+++ samples/S3_SimpleDB_SNS_SQS_Demo/AWSiOSDemo/Classes/S3/S3ResponseHandler.m
@@ -62,7 +62,7 @@
     bytesIn.text = [NSString stringWithFormat:@"%d", total];
 }
 
--(void)request:(AmazonServiceRequest *)request didSendData:(NSInteger)bytesWritten totalBytesWritten:(NSInteger)totalBytesWritten totalBytesExpectedToWrite:(NSInteger)totalBytesExpectedToWrite
+-(void)request:(AmazonServiceRequest *)request didSendData:(long long)bytesWritten totalBytesWritten:(long long)totalBytesWritten totalBytesExpectedToWrite:(long long)totalBytesExpectedToWrite
 {
     NSLog(@"didSendData");
     int total = [bytesOut.text intValue];
