+++ src/Amazon.Runtime/AmazonRequestDelegate.m
@@ -57,7 +57,7 @@
     AMZLogDebug(@"didReceiveData");
 }
 
--(void)request:(AmazonServiceRequest *)request didSendData:(NSInteger)bytesWritten totalBytesWritten:(NSInteger)totalBytesWritten totalBytesExpectedToWrite:(NSInteger)totalBytesExpectedToWrite
+-(void)request:(AmazonServiceRequest *)request didSendData:(long long)bytesWritten totalBytesWritten:(long long)totalBytesWritten totalBytesExpectedToWrite:(long long)totalBytesExpectedToWrite
 {
     AMZLogDebug(@"didSendData");
 }
