+++ src/include/AmazonRequestDelegate.h
@@ -32,7 +32,7 @@
 -(void)request:(AmazonServiceRequest *)request didReceiveResponse:(NSURLResponse *)response;
 -(void)request:(AmazonServiceRequest *)request didCompleteWithResponse:(AmazonServiceResponse *)response;
 -(void)request:(AmazonServiceRequest *)request didReceiveData:(NSData *)data;
--(void)request:(AmazonServiceRequest *)request didSendData:(NSInteger)bytesWritten totalBytesWritten:(NSInteger)totalBytesWritten totalBytesExpectedToWrite:(NSInteger)totalBytesExpectedToWrite;
+-(void)request:(AmazonServiceRequest *)request didSendData:(long long)bytesWritten totalBytesWritten:(long long)totalBytesWritten totalBytesExpectedToWrite:(long long)totalBytesExpectedToWrite;
 -(void)request:(AmazonServiceRequest *)request didFailWithError:(NSError *)error;
 -(void)request:(AmazonServiceRequest *)request didFailWithServiceException:(NSException *)exception;
